requirements.txt
yarl<1.2
discord.py
